more moving validators around
